Add rib rename action
