+++ android/tooling/rib-intellij-plugin/src/main/java/com/badoo/ribs/plugin/generator/dialog/GenerateRibDialog.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.badoo.ribs.plugin.action.dialog.GenerateRibDialog">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.badoo.ribs.plugin.generator.dialog.GenerateRibDialog">
   <grid id="27dc6" binding="contentPanel" layout-manager="GridLayoutManager" row-count="3" column-count="3" same-size-horizontally="false" same-size-vertically="false" hgap="0" vgap="0">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
