+++ android/tooling/rib-intellij-plugin/src/main/java/com/badoo/ribs/plugin/generator/dialog/TemplateComboBoxModel.kt
@@ -1,4 +1,4 @@
-package com.badoo.ribs.plugin.action.dialog
+package com.badoo.ribs.plugin.generator.dialog
 
 import com.badoo.ribs.plugin.template.Template
 import javax.swing.DefaultComboBoxModel
