+++ android/tooling/rib-intellij-plugin/src/main/java/com/badoo/ribs/plugin/icons/RIBIconProvider.kt
@@ -6,9 +6,12 @@ import com.intellij.openapi.util.IconLoader
 import com.intellij.openapi.util.Key
 import com.intellij.psi.PsiDirectory
 import com.intellij.psi.PsiElement
+import com.intellij.psi.util.CachedValueProvider
+import com.intellij.psi.util.CachedValuesManager
 import org.jetbrains.kotlin.psi.KtClass
 import org.jetbrains.kotlin.psi.KtSimpleNameExpression
 import org.jetbrains.kotlin.psi.KtVisitorVoid
+import org.jetbrains.kotlin.utils.addToStdlib.firstNotNullResult
 import javax.swing.Icon
 
 class RIBIconProvider: IconProvider() {
@@ -24,16 +27,16 @@ class RIBIconProvider: IconProvider() {
     }
 
     private fun PsiElement.hasRib(): Boolean {
-        // Check the cache for result
-        getUserData(ribKey)?.let {
-            return it
-        }
+//        // Check the cache for result
+//        getUserData(IS_RIB_KEY)?.let {
+//            return it
+//        }
 
         if (DumbService.isDumb(project)) return false
 
         val childFiles = children.filter { it !is PsiDirectory }
-        val hasRib = childFiles.any {
-            var isRib = false
+        val ribName = childFiles.firstNotNullResult {
+            var ribName: String? = null
             it.acceptChildren(object : KtVisitorVoid() {
                 /**
                  * Search for the interface with Rib interface as super
@@ -42,17 +45,24 @@ class RIBIconProvider: IconProvider() {
                     super.visitClass(klass)
                     if (!klass.isInterface()) return
 
-                    isRib = klass.superTypeListEntries.mapNotNull { it.typeAsUserType?.referenceExpression }
+                    val isRib = klass.superTypeListEntries
+                        .mapNotNull { it.typeAsUserType?.referenceExpression }
                         .any { it.isRib() }
+
+                    ribName = if (isRib) klass.name else null
                 }
             })
-            isRib
+            ribName
         }
 
-        // Cache the result
-        putUserData(ribKey, hasRib)
+        CachedValuesManager.getManager(project)
+            .createCachedValue { CachedValueProvider.Result(ribName == null, children) }
 
-        return hasRib
+        // Cache the result
+        putUserData(IS_RIB_KEY, ribName != null)
+        putUserData(RIB_NAME_KEY, ribName)
+
+        return ribName != null
     }
 
     fun KtSimpleNameExpression.isRib(): Boolean {
@@ -64,6 +74,7 @@ class RIBIconProvider: IconProvider() {
     companion object {
         private const val RIB_FQ_NAME = "com.badoo.ribs.core.Rib"
 
-        private val ribKey = Key<Boolean>("com.badoo.ribs.isRib")
+        val IS_RIB_KEY = Key<Boolean>("com.badoo.ribs.isRib")
+        val RIB_NAME_KEY = Key<String>("com.badoo.ribs.ribName")
     }
 }
